feat(home-feed): replace reflection producer with activity-log roll-up [JARVIS-512]
Pivot the LLM-driven producer from "reflect on relationship state from
nothing" → "consolidate recent action items into digests/threads."
Same skeleton, same cadence mechanism (30 min), new input and new
intent.

Rename reflection-producer.ts → rollup-producer.ts (git rename
tracked so blame history is preserved). The new producer:

  - Reads recent `action` items from readHomeFeed() via a new
    defaultLoadRecentActions helper (sorted newest-first, capped at
    MAX_ACTIONS_IN_PROMPT = 30 for token budget). Non-action items
    are excluded — the roll-up's input is the raw activity log, not
    existing consolidations.
  - Still loads relationship state for context, but only as
    secondary framing (explicitly instructed in the prompt NOT to
    invent roll-ups from facts alone).
  - Narrows the write_feed_items tool schema to type: digest | thread
    only. No nudges, no actions — actions come from background jobs,
    not the LLM.
  - New `no_actions` skip reason: when the activity log is empty,
    the producer short-circuits before the provider call. The
    scheduler treats this like no_provider — does NOT advance the
    cooldown gate, so the next tick retries as soon as new actions
    land instead of waiting the full 30-minute window. Real LLM
    attempts (success/empty_items/malformed_output/provider_error)
    still advance the gate to preserve backoff.

feed-scheduler.ts renamed symbols to match:
  reflectionRunner → rollupRunner
  reflectionRan → rollupRan
  REFLECTION_INTERVAL_MS → ROLLUP_INTERVAL_MS
  lastReflectionAt → lastRollupAt

Cadence stays at 30 min in this PR. The drop to 120 min + on-visit
refresh in home-feed-routes.ts is the next step in JARVIS-512.

Tests migrated from reflection-producer.test.ts → rollup-producer.
test.ts with all existing coverage plus new cases:
  - no_actions short-circuits before the provider call
  - action items are serialized into the user prompt
  - nudge/action types are rejected at coercion time even if the
    model bypasses the tool schema

feed-scheduler.test.ts adds a no_actions cooldown test alongside
the existing no_provider test.
